#patch Fixed plugin "crash" none paper software servers  (#56)
* Bump version from 1.3.11 to 1.3.11-SNAPSHOT0

* Fixed VersionUtils.teleport on none paper software servers

* Bump version from 1.3.11-SNAPSHOT0 to 1.3.11-SNAPSHOT1

* Fixed CommandArgument could have multiple permissions while only the first one was checked

* Removed depreciated class User usage

* Fixed MaterialUtils.getMat did not force default material on null

* Fixed some sonarcloud code bugs

* Bump version from 1.3.11-SNAPSHOT1 to 1.3.11-SNAPSHOT2

* Fixed Security Hotspots

* Bump version from 1.3.11-SNAPSHOT2 to 1.3.11-SNAPSHOT3

* Revert MaterialUtils

* Bump version from 1.3.11-SNAPSHOT3 to 1.3.11-SNAPSHOT4

* Optimized Command Permission check

* Bump version from 1.3.11-SNAPSHOT4 to 1.3.11-SNAPSHOT5

---------

+++ MiniGamesBox Classic/src/main/java/plugily/projects/minigamesbox/classic/utils/helper/BukkitHelper.java
@@ -78,7 +78,7 @@ public class BukkitHelper {
    * @return serialized number
    */
   public int serializeInt(int i) {
-    return (i % 9) == 0 ? i : (int) ((Math.ceil(i / 9) * 9) + 9);
+    return (i % 9) == 0 ? i : (int) ((Math.ceil((double) i / 9) * 9) + 9);
   }
 
   @SuppressWarnings("deprecation")
